refactor: import하는 User 타입 변경
