+++ src/components/molecules/reply/reply-header.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import { Avatar } from '@/components/atoms/avatar';
-import { User } from '@/types';
+import { User } from '@/stores/useAuthStore';
 
 type ReplyHeaderProps = {
-  writer: Omit<User, 'email'>;
+  writer: Pick<User, 'userId' | 'nickname' | 'profileImage'>;
   createdAt: string;
   replyId: number;
 };
@@ -28,13 +28,13 @@ export const ReplyHeader = ({
     <header className="flex justify-between">
       <div className="flex gap-3 items-center">
         <Avatar
-          imageSrc={writer.profileImage}
-          fallback={writer.name[0]}
+          imageSrc={writer.profileImage ?? 'https://github.com/shadcn.png'}
+          fallback={writer.nickname?.[0] ?? ''}
           className="w-10 h-10 cursor-pointer"
           onClick={() => {}}
         />
         <div className="flex flex-col">
-          <div>{writer.name}</div>
+          <div>{writer.nickname}</div>
           <div>{createdAt}</div>
         </div>
       </div>
